+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![allow(clippy::into_iter_on_ref)]
 
+mod aggregated_queries;
 mod cli;
 mod config;
 mod context;
@@ -22,9 +23,10 @@ use std::{
     ops::Deref,
     path::{Path, PathBuf},
     process,
-    sync::{Arc, Mutex, PoisonError},
+    sync::{Mutex, PoisonError},
 };
 
+use aggregated_queries::AggregatedQueries;
 pub use cli::bootstrap_cli;
 pub use config::{Args, ArgsBuilder, Config, ConfigBuilder, RuleConfiguration};
 use context::PendingFix;
@@ -43,7 +45,7 @@ pub use rule_tester::{
     RuleTestValid, RuleTester, RuleTests,
 };
 pub use slice::MutRopeOrSlice;
-use tree_sitter::{Query, Tree};
+use tree_sitter::Tree;
 use tree_sitter_grep::{
     get_matches, get_parser, streaming_iterator::StreamingIterator, tree_sitter::QueryMatch,
     Parseable, RopeOrSlice, SupportedLanguage,
@@ -56,8 +58,6 @@ pub extern crate serde_yaml;
 pub extern crate tokio;
 pub extern crate tree_sitter_grep;
 pub use tree_sitter_grep::{ropey, tree_sitter};
-
-use crate::rule::ResolvedMatchBy;
 
 pub fn run_and_output(config: Config) {
     let violations = run(&config);
@@ -72,28 +72,24 @@ pub fn run_and_output(config: Config) {
 
 const MAX_FIX_ITERATIONS: usize = 10;
 
-#[allow(clippy::too_many_arguments)]
 fn run_per_file<'a>(
     file_run_context: FileRunContext<'a>,
-    aggregated_queries: &'a AggregatedQueries<'a>,
     mut on_found_violations: impl FnMut(Vec<ViolationWithContext>),
     mut on_found_pending_fixes: impl FnMut(Vec<PendingFix>, &InstantiatedRule),
-    query: &'a Arc<Query>,
-    instantiated_rules: &'a [InstantiatedRule],
 ) {
     let mut instantiated_per_file_rules: HashMap<RuleName, Box<dyn RuleInstancePerFile<'a> + 'a>> =
         Default::default();
     get_matches(
         file_run_context.language.language(),
         file_run_context.file_contents,
-        query,
+        file_run_context.query,
         Some(file_run_context.tree),
     )
     .for_each(|query_match| {
         run_match(
             file_run_context,
             query_match,
-            aggregated_queries,
+            file_run_context.aggregated_queries,
             &mut instantiated_per_file_rules,
             &mut on_found_violations,
             |pending_fixes, instantiated_rule| {
@@ -102,18 +98,22 @@ fn run_per_file<'a>(
         );
     });
 
-    for (&instantiated_rule_index, &rule_listener_index) in
-        &aggregated_queries.instantiated_rule_root_exit_rule_listener_indices
+    for (&instantiated_rule_index, &rule_listener_index) in &file_run_context
+        .aggregated_queries
+        .instantiated_rule_root_exit_rule_listener_indices
     {
         run_single_on_query_match_callback(
             file_run_context,
-            &instantiated_rules[instantiated_rule_index],
+            &file_run_context.instantiated_rules[instantiated_rule_index],
             &mut instantiated_per_file_rules,
             rule_listener_index,
             file_run_context.tree.root_node().into(),
             &mut on_found_violations,
             |pending_fixes| {
-                on_found_pending_fixes(pending_fixes, &instantiated_rules[instantiated_rule_index])
+                on_found_pending_fixes(
+                    pending_fixes,
+                    &file_run_context.instantiated_rules[instantiated_rule_index],
+                )
             },
         );
     }
@@ -129,8 +129,16 @@ pub fn run(config: &Config) -> Vec<ViolationWithContext> {
         tree_sitter_grep_args,
         |dir_entry, language, file_contents, tree, query| {
             run_per_file(
-                FileRunContext::new(dir_entry.path(), file_contents, tree, config, language),
-                &aggregated_queries,
+                FileRunContext::new(
+                    dir_entry.path(),
+                    file_contents,
+                    tree,
+                    config,
+                    language,
+                    &aggregated_queries,
+                    query,
+                    &instantiated_rules,
+                ),
                 |violations| {
                     all_violations
                         .lock()
@@ -148,8 +156,6 @@ pub fn run(config: &Config) -> Vec<ViolationWithContext> {
                         language,
                     )
                 },
-                query,
-                &instantiated_rules,
             );
         },
     )
@@ -251,7 +257,6 @@ fn run_match<'a, 'b>(
     }
 }
 
-#[allow(clippy::too_many_arguments)]
 fn run_single_on_query_match_callback<'a, 'b>(
     file_run_context: FileRunContext<'a>,
     instantiated_rule: &'a InstantiatedRule,
@@ -313,8 +318,20 @@ fn run_fixing_loop<'a>(
             .parse(&mut get_parser(language.language()), None)
             .unwrap();
         run_per_file(
-            FileRunContext::new(path, &file_contents, &tree, config, language),
-            aggregated_queries,
+            FileRunContext::new(
+                path,
+                &file_contents,
+                &tree,
+                config,
+                language,
+                aggregated_queries,
+                &aggregated_queries
+                    .per_language
+                    .get(&language)
+                    .unwrap()
+                    .query,
+                instantiated_rules,
+            ),
             |reported_violations| {
                 violations.extend(reported_violations);
             },
@@ -324,12 +341,6 @@ fn run_fixing_loop<'a>(
                     .or_default()
                     .extend(fixes);
             },
-            &aggregated_queries
-                .per_language
-                .get(&language)
-                .unwrap()
-                .query,
-            instantiated_rules,
         );
         if pending_fixes.is_empty() {
             break;
@@ -363,20 +374,26 @@ pub fn run_for_slice<'a>(
         Cow::Borrowed,
     );
     run_per_file(
-        FileRunContext::new(path, file_contents, &tree, &config, language),
-        &aggregated_queries,
+        FileRunContext::new(
+            path,
+            file_contents,
+            &tree,
+            &config,
+            language,
+            &aggregated_queries,
+            &aggregated_queries
+                .per_language
+                .get(&language)
+                .unwrap()
+                .query,
+            &instantiated_rules,
+        ),
         |reported_violations| {
             violations.lock().unwrap().extend(reported_violations);
         },
         |_, _| {
             panic!("Expected no fixes");
         },
-        &aggregated_queries
-            .per_language
-            .get(&language)
-            .unwrap()
-            .query,
-        &instantiated_rules,
     );
     violations.into_inner().unwrap()
 }
@@ -408,8 +425,20 @@ pub fn run_fixing_for_slice<'a>(
     let violations: Mutex<Vec<ViolationWithContext>> = Default::default();
     let pending_fixes: Mutex<HashMap<RuleName, Vec<PendingFix>>> = Default::default();
     run_per_file(
-        FileRunContext::new(path, &file_contents, &tree, &config, language),
-        &aggregated_queries,
+        FileRunContext::new(
+            path,
+            &file_contents,
+            &tree,
+            &config,
+            language,
+            &aggregated_queries,
+            &aggregated_queries
+                .per_language
+                .get(&language)
+                .unwrap()
+                .query,
+            &instantiated_rules,
+        ),
         |reported_violations| {
             violations.lock().unwrap().extend(reported_violations);
         },
@@ -421,12 +450,6 @@ pub fn run_fixing_for_slice<'a>(
                 .or_default()
                 .extend(fixes);
         },
-        &aggregated_queries
-            .per_language
-            .get(&language)
-            .unwrap()
-            .query,
-        &instantiated_rules,
     );
     let mut violations = violations.into_inner().unwrap();
     let pending_fixes = pending_fixes.into_inner().unwrap();
@@ -586,128 +609,5 @@ impl PerFilePendingFixes {
             pending_fixes: Default::default(),
             language,
         }
-    }
-}
-
-type RuleIndex = usize;
-type RuleListenerIndex = usize;
-type CaptureIndexIfPerCapture = Option<u32>;
-
-struct AggregatedQueriesPerLanguage {
-    pattern_index_lookup: Vec<(RuleIndex, RuleListenerIndex, CaptureIndexIfPerCapture)>,
-    query: Arc<Query>,
-    #[allow(dead_code)]
-    query_text: String,
-}
-
-#[derive(Default)]
-struct AggregatedQueriesPerLanguageBuilder {
-    pattern_index_lookup: Vec<(RuleIndex, RuleListenerIndex, CaptureIndexIfPerCapture)>,
-    query_text: String,
-}
-
-impl AggregatedQueriesPerLanguageBuilder {
-    pub fn build(self, language: SupportedLanguage) -> AggregatedQueriesPerLanguage {
-        let Self {
-            pattern_index_lookup,
-            query_text,
-        } = self;
-        let query = Arc::new(Query::new(language.language(), &query_text).unwrap());
-        assert!(query.pattern_count() == pattern_index_lookup.len());
-        AggregatedQueriesPerLanguage {
-            pattern_index_lookup,
-            query,
-            query_text,
-        }
-    }
-}
-
-struct AggregatedQueries<'a> {
-    instantiated_rules: &'a [InstantiatedRule],
-    per_language: HashMap<SupportedLanguage, AggregatedQueriesPerLanguage>,
-    pub instantiated_rule_root_exit_rule_listener_indices: HashMap<usize, usize>,
-}
-
-impl<'a> AggregatedQueries<'a> {
-    pub fn new(instantiated_rules: &'a [InstantiatedRule]) -> Self {
-        let mut per_language: HashMap<SupportedLanguage, AggregatedQueriesPerLanguageBuilder> =
-            Default::default();
-        let mut instantiated_rule_root_exit_rule_listener_indices: HashMap<usize, usize> =
-            Default::default();
-        for (rule_index, instantiated_rule) in instantiated_rules.into_iter().enumerate() {
-            for &language in &instantiated_rule.meta.languages {
-                let per_language_builder = per_language.entry(language).or_default();
-                for (rule_listener_index, rule_listener_query) in instantiated_rule
-                    .rule_instance
-                    .listener_queries()
-                    .iter()
-                    .enumerate()
-                    .filter_map(|(rule_listener_index, rule_listener_query)| {
-                        if rule_listener_query.query == ROOT_EXIT {
-                            instantiated_rule_root_exit_rule_listener_indices
-                                .insert(rule_index, rule_listener_index);
-
-                            return None;
-                        }
-
-                        Some((
-                            rule_listener_index,
-                            rule_listener_query.resolve(language.language()),
-                        ))
-                    })
-                {
-                    let capture_index_if_per_capture: CaptureIndexIfPerCapture =
-                        match &rule_listener_query.match_by {
-                            ResolvedMatchBy::PerCapture { capture_index } => Some(*capture_index),
-                            _ => None,
-                        };
-
-                    for _ in 0..rule_listener_query.query.pattern_count() {
-                        per_language_builder.pattern_index_lookup.push((
-                            rule_index,
-                            rule_listener_index,
-                            capture_index_if_per_capture,
-                        ));
-                    }
-                    per_language_builder
-                        .query_text
-                        .push_str(&rule_listener_query.query_text);
-                    per_language_builder.query_text.push_str("\n\n");
-                }
-            }
-        }
-
-        Self {
-            instantiated_rules,
-            per_language: per_language
-                .into_iter()
-                .map(|(language, per_language_value)| {
-                    (language, per_language_value.build(language))
-                })
-                .collect(),
-            instantiated_rule_root_exit_rule_listener_indices,
-        }
-    }
-
-    pub fn get_rule_and_listener_index_and_capture_index(
-        &self,
-        language: SupportedLanguage,
-        pattern_index: usize,
-    ) -> (&'a InstantiatedRule, usize, CaptureIndexIfPerCapture) {
-        let (rule_index, rule_listener_index, capture_index_if_per_capture) = self
-            .per_language
-            .get(&language)
-            .unwrap()
-            .pattern_index_lookup[pattern_index];
-        let instantiated_rule = &self.instantiated_rules[rule_index];
-        (
-            instantiated_rule,
-            rule_listener_index,
-            capture_index_if_per_capture,
-        )
-    }
-
-    pub fn get_query_for_language(&self, language: SupportedLanguage) -> Arc<Query> {
-        self.per_language.get(&language).unwrap().query.clone()
     }
 }
